Add fallback background-color for .masthead
In IE9, .masthead was displaying as blank. The background was only styled with gradients. I've added a fallback background color for those browsers that don't support gradients. I chose the darkest of the gradient colors used as it has the highest contrast rating for any size text (http://leaverou.github.com/contrast-ratio/#%23fff-on-%231a274e).