Added legend element to checklist
The legend explains to screen-readers what the group of fields is about. Especially important for groups of checkboxes or radio buttons, as the legend will be the question that the checkboxes or radio buttons are answers for. "The first element inside the fieldset must be a legend element, which provides a label or description for the group" - http://www.w3.org/TR/2012/NOTE-WCAG20-TECHS-20120103/H71