update heading nesting
update level 4 headings to level 3 as that would be the expected nesting order after a level 2 heading.

update .excerpt-list__title class to have the same font-size as the h4 selector, to maintain visual consistency.

This commit also contains adding a closing </li> tag for the category posts, as closing </li>s are used for other lists on the site, and this will fix HTML validation errors that are being reported by the browser.