Remove role="navigation" from example
There is some contention about whether or not `<nav>` is semantically the same as `<nav role="navigation">`, but the October 2014 version of the HTML spec seems to indicate it shouldn't be set:

https://www.w3.org/TR/html5/sections.html#the-nav-element

Even if `<nav>` implicitly having the "navigation" landmark isn't fully implemented by all screen reader, I would think this documentation should be aspirational, showing how things will be. It also contradicts the important, and I think generally hard to grasp the idea, that redundant roles should not be used. It is hard to grasp due to a proliferation of incorrect examples.

That being said having `role="navigation"` still may be practical and a good example, but if so my feeling is that should be explained.