Fix Safari focus outline bug (#950) (#1032)
* Remove text-decoration on focus/hover only for anchor elements with no
  class applied

*** More Details ***

For some reason that Google didn't reveal to me,

```
a:focus { text-decoration: none; }
```

prevents the outline from rendering in Safari. However, if we scope that
style to only anchor elements without a class applied (which is what the
code in `_links.scss` was seemingly intended to do), then our focus
issue goes away.

I'm not sure of the exact intent of the CSS in
[this commit](https://github.com/a11yproject/a11yproject.com/commit/ba9c4c04d3bd3ecdeca97ea7e0b3f54e2700eae7#diff-10dc230e000876829e9e47ff5fa31d26R8-R15),
but it looks like a global style intended primarily for inline text
links; a hover/focus style different from the default underline
applied by the user agent.

I also noticed a comment at the top of `_link.scss` that seemed
relevant:

> // Initial styles applied to anchor elements without a class applied
> // to them.

This suggests that all the styles in the file should apply to links
without a class attribute, but the related CSS (`&[href]:not([class])`)
was previously commented out. Now links with no class applied (e.g.
inline text links) will have their default text-decoration removed on
focus and hover, but links with classes (e.g. .c-sponsor__link) will
not.