Adding myself to authors file
Seems to be in alphabetical order by first name if I caught the pattern correctly. Of course please fix if not.
Also, I don't have a direct URL for my photo. I looked into the "githubusercontent" and "pbs.twimg.com" paths others used but I can't figure out how to leverage those. If I can email a photo to put on the A11Y Project servers that'd be great - thanks!